requirements.txt
torch~=2.1.2
pillow~=9.3.0
Flask~=3.0.0
transformers~=4.36.2
gradio
requests~=2.28.1
numpy~=1.24.1